requirements.txt
Django
gunicorn
python-telegram-bot
python-dotenv
